Split tiles to story/examples
