+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
-members = ["crates/app", "crates/story", "crates/ui"]
+members = ["crates/story", "crates/ui"]
 
-default-members = ["crates/app"]
+default-members = ["crates/story"]
 resolver = "2"
 
 [workspace.dependencies]
@@ -11,6 +11,7 @@ log = "0.4"
 serde = "1.0.203"
 serde_json = "1"
 smallvec = "1"
+rust-embed = "8.5.0"
 story = { path = "crates/story" }
 ui = { path = "crates/ui" }
 windows = { version = "0.58.0", features = [
