--- crates/app/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "gpui-app"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-gpui.workspace = true
-anyhow.workspace = true
-rust-embed = "8"
-log.workspace = true
-ui.workspace = true
-story.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-rand = "0.8"
-
-[lints]
-workspace = true
