+++ crates/story/Cargo.toml
@@ -13,6 +13,9 @@ chrono = "0.4"
 serde = "1"
 serde_json = "1"
 unindent = "0.2.3"
+anyhow.workspace = true
+rust-embed.workspace = true
+rand = "0.8"
 
 [lints]
 workspace = true
