+++ crates/story/src/assets.rs
