+++ crates/ui/Cargo.toml
@@ -21,7 +21,7 @@ resvg = { version = "0.44.0", default-features = false, features = [
   "system-fonts",
   "text",
 ] }
-rust-embed = "8.5.0"
+rust-embed.workspace = true
 rust-i18n = "3"
 serde = "1.0.203"
 serde_json = "1"
